Fix source not being removed when peer is removed from Torrent
Previously, a new Observable instance is created whenever
getDownloadRateObservable() or getUploadRateObservable() is called
which results in trying to remove the new Observable instance from
the set of sources instead of the original Observable instance that was
added to it.

Now, the Observables are created only once in Peer and the same
instances are returned every time.